Extract handler for non-shifted special keys
